refactor: 패키지 이동
Contract 인터페이스들을 contract라는 새로운 패키지로 이동
