+++ app/src/main/java/woowacourse/movie/adapter/MovieListAdapter.kt
@@ -6,7 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseAdapter
 import woowacourse.movie.R
-import woowacourse.movie.domain.model.Movie
+import woowacourse.movie.model.Movie
 import java.time.format.DateTimeFormatter
 
 class MovieListAdapter(
