+++ app/src/main/java/woowacourse/movie/contract/MovieReservationContract.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.contract
 
 interface MovieReservationContract {
     interface View {
