+++ app/src/main/java/woowacourse/movie/model/MovieData.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.model
 
 import woowacourse.movie.R
 import java.time.LocalDateTime
