+++ app/src/main/java/woowacourse/movie/model/Ticket.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain.model
+package woowacourse.movie.model
 
 import java.io.Serializable
