+++ app/src/main/java/woowacourse/movie/presenter/MovieReservationPresenter.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.presenter
 
-import woowacourse.movie.domain.MovieReservationContract
-import woowacourse.movie.domain.model.MovieReservationModel
+import woowacourse.movie.contract.MovieReservationContract
+import woowacourse.movie.model.MovieReservationModel
 
 class MovieReservationPresenter(
     private val view: MovieReservationContract.View,
