+++ app/src/main/java/woowacourse/movie/view/MovieReservationActivity.kt
@@ -9,9 +9,9 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.movie.R
-import woowacourse.movie.domain.MovieReservationContract
-import woowacourse.movie.domain.model.Movie
-import woowacourse.movie.domain.model.Ticket
+import woowacourse.movie.contract.MovieReservationContract
+import woowacourse.movie.model.Movie
+import woowacourse.movie.model.Ticket
 import woowacourse.movie.presenter.MovieReservationPresenter
 import java.io.Serializable
 import java.time.format.DateTimeFormatter
