+++ app/src/main/java/woowacourse/movie/view/MovieReservationCompleteActivity.kt
@@ -5,7 +5,7 @@ import android.os.Bundle
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.movie.R
-import woowacourse.movie.domain.model.Ticket
+import woowacourse.movie.model.Ticket
 
 class MovieReservationCompleteActivity : AppCompatActivity() {
     private lateinit var ticketTitle: TextView
